refactor(api): co-locate eln error types in an errors module
The eln error types lived where they were produced: ElnError and
ElnErrorCode in archive.ts (the parser was the only validator), and
ElnImportError in import.service.ts.

This is prep work: validation is about to stop being archive-only —
different translators (SciLog, openBIS) have different validation needs
and will each produce their own errors — and the translator registry
will need to throw ElnImportError. None of those can import the error
vocabulary from the parser or the service without an awkward
dependency (a cycle, in the registry's case). Move the vocabulary into
a neutral errors module so the upcoming translator-side validation and
registry can depend on it cleanly.

The parse-result types (ElnParse*) stay in archive.ts as they are
specific to ElnArchive.parse. Behaviour-preserving; no new consumers
yet.
